+++ app/myapi/routes.py
@@ -63,95 +63,5 @@ def resources_init(myresources):
 # === Load each API resource ===
 resources_init(resources.resources)
 
-#################################################################
-#################################################################
-#################################################################
-#################################################################
-
-##################################
-# Rethinkdb connection
-import os
-import rethinkdb as r
-RDB_HOST = "db"
-RDB_PORT = os.environ.get('DB_PORT_28015_TCP_PORT') or 28015
-DB = 'webapp'
-TABLE = 'objtest'
-FIELD = 'latest_timestamp'
-params = {"host":RDB_HOST, "port":RDB_PORT}
-r.connect(**params).repl()
-base = r.db(DB)
-
-##########################################
-# Read model template
-import glob, json
-JSONS_PATH = 'jsonmodels'
-JSONS_EXT = 'json'
-for fileschema in glob.glob(os.path.join(JSONS_PATH, "*") + "." + JSONS_EXT):
-    data = {}
-    with open(fileschema) as f:
-        data = json.load(f)
-    print(data)
-    break
-
-##########################################
-# Build current model
-import re
-myre = re.compile('^http[s]?://[^\.]+\..{2,}')
-from flask.ext.restful import fields
-
-def marshal_type(obj):
-#http://flask-restful-cn.readthedocs.org/en/0.3.4/api.html#module-fields
-    mytype = fields.Raw
-    if isinstance(obj, str) or isinstance(obj, unicode):
-        if myre.search(obj):
-            mytype = fields.Url()#endpoint=obj, absolute=True)
-        else:
-            mytype = fields.String
-    if isinstance(obj, int):
-        mytype = fields.Integer
-    return mytype
-
-def convert_to_marshal(data):
-    ## recursive?
-    mymarshal = {}
-    # Case of dict
-    for key, obj in data.iteritems():
-        mymarshal[key] = marshal_type(obj)
-    # Case of lists?
-    print(mymarshal)
-    return mymarshal
-
-schema = convert_to_marshal(data)
-
-##########################################
-# Build resource
-from flask.ext.restful import Resource, marshal
-from flask import Flask, request
-
-class Test(Resource):
-    def post(self):
-        json_data = request.get_json(force=True) # this issues Bad request
-        print("Raw json", json_data)
-        marshal_data = marshal(json_data, schema, envelope='data')
-        print("Interpreted", marshal_data)
-
-        # Rethinkdb base query
-        query = base.table(TABLE)
-
-        # CREATE?
-        #base.table_create(TABLE).run()
-
-        # Execute the insert
-        dbout = query.insert(marshal_data['data']).run()
-
-        # GET THE ID
-        myid = dbout['generated_keys'].pop()
-
-# // TO FIX:
-# redirect to GET method of this same endpoint, instead
-        # Recover the element to check we are done
-        document = query.get(myid).run()
-        document.pop('id')
-        return document
-
+from myapi.experiments import Test
 api.add_resource(Test, '/objtest')
